-Adding the new configuration widget -Loading / saving functionalities restored -Adding the new bank selection system -Computer keyboard can now trigger MIDI notes
git-svn-id: https://svn.code.sf.net/p/fb01editor/code@28 48d7863b-f1c1-4a9b-9dfa-e6142b70e2fc

+++ core/edit.cpp
@@ -53,7 +53,7 @@ bool Edit::Enregistrer(FILE * fichier)
     return true;
 }
 
-bool Edit::Charger(FILE * fichier, const int version)
+bool Edit::Charger(FILE * fichier, const short version)
 {
     uchar sauv[nbParam];
     if (fread(sauv, nbParam, 1, fichier))
